reorganize directory tests
